выбор типа вопроса при его добавлении
